configure.ac: Require liboil.
Original commit message from CVS:
* configure.ac: Require liboil.
* gst/videotestsrc/gstvideotestsrc.c: Fix up liboil calls, add
a few more.
* gst/videotestsrc/videotestsrc.c:
* gst/videotestsrc/videotestsrc.h:

+++ gst/videotestsrc/videotestsrc.h
@@ -20,9 +20,11 @@
 #ifndef __VIDEOTESTSRC_H__
 #define __VIDEOTESTSRC_H__
 
+#include <stdint.h>
+
 struct vts_color_struct {
-	int Y, U, V;
-	int R, G, B;
+	uint8_t Y, U, V;
+	uint8_t R, G, B;
 };
 
 typedef struct paintinfo_struct paintinfo;
